Updated edges, script and some documentation
