+++ docs/basics/paper.md
@@ -112,7 +112,7 @@ Now that you have an idea of building a Paper Properties object, let's look at
 initializing a Paper instance.
 
 ```js
-import { Paper, Rectangle, TextEdge } from 'psiagram';
+import { Paper, Rectangle, TextLine } from 'psiagram';
 import { Grid } from 'psiagram-plugin-grid';
 import { MouseEvents } from 'psiagram-plugin-mouse-events';
 import { ManhattanRouting } from 'psiagram-plugin-routing';
@@ -147,7 +147,7 @@ const myPaper = new Paper({
         properties: { title: 'Edge 1' },
       },
     ],
-    edgeComponentMap: { 'text-edge': TextEdge },
+    edgeComponentMap: { 'text-edge': TextLine },
   },
 });
 ```
