+++ docs/in-depth/README.md
@@ -1,3 +1,9 @@
 # In Depth
 
-🚧 Section is coming very soon. 🚧
+Now that you've touched on the basics, you are ready to get into some more
+advanced Psiagram functionality.
+
+- [Custom Nodes](custom-nodes.md)
+- [Custom Edges](custom-edges.md)
+- [Active Item](active-item.md)
+- [Events](events.md)
