+++ docs/in-depth/events.md
@@ -1,0 +1,3 @@
+# Edges
+
+🚧 Section is coming very soon. 🚧
